Made flash utilities executable. Moved creation of default /etc/waggle/devices file to the waggle-communications service. Made removal of registration key dependent on its existance.

+++ nc-wag-os/waggled/Communications/communications_main.py
@@ -138,7 +138,8 @@ def get_certificates():
                 continue
         else:
             logger.info("All certificate files found.")
-            os.remove(REGISTRATION_KEY)
+            if os.path.isfile(REGISTRATION_KEY):
+              os.remove(REGISTRATION_KEY)
             break
             
         # make sure certficate files exist.
